UCS/NVTX: move profile range fxns to profile_on
