+++ src/ucs/profile/profile.h
@@ -17,6 +17,4 @@
 #  include "profile_off.h"
 #endif
 
-#include "profile_range.h"
-
 #endif
