+++ src/ucs/profile/profile_on.h
@@ -1,4 +1,5 @@
 /**
+* Copyright (C) NVIDIA Corporation. 2021.  ALL RIGHTS RESERVED.
 * Copyright (C) Mellanox Technologies Ltd. 2001-2018.  ALL RIGHTS RESERVED.
 *
 * See file LICENSE for terms.
@@ -12,6 +13,7 @@
 #include <ucs/sys/compiler_def.h>
 #include <ucs/sys/preprocessor.h>
 #include <ucs/config/global_opts.h>
+#include <stdint.h>
 
 
 BEGIN_C_DECLS
@@ -282,6 +284,53 @@ void ucs_profile_record(ucs_profile_type_t type, const char *name,
  */
 void ucs_profile_reset_locations();
 
+
+/*
+ * Start a range trace on an arbitrary event in a potentially nested fashion.
+ * A range tracing can be started in a function and potentially end in an
+ * another function or a recursive invocation of the same function.
+ * A unique ID is returned to stop tracing the event.
+ *
+ * @param [in]     format      String name for the range.
+ *
+ * @return ID to be used to stop tracing a range
+ */
+uint64_t ucs_profile_range_start(const char *format, ...);
+
+
+/*
+ * Stop range trace.
+ *
+ * @param [in]     id          id that was returned from range start.
+ *
+ */
+void ucs_profile_range_stop(uint64_t id);
+
+
+/*
+ * Add a marker on trace profiles.
+ *
+ * @param [in]     format      String name for the marker.
+ *
+ */
+void ucs_profile_range_add_marker(const char *format, ...);
+
+
+/*
+ * Start a range trace in a non-nested fashion. Range tracing must start and end
+ * in the same function.
+ *
+ * @param [in]     format      String name for the marker.
+ *
+ */
+void ucs_profile_range_push(const char *format, ...);
+
+
+/*
+ * Stop a range trace in a non-nested fashion.
+ *
+ */
+void ucs_profile_range_pop();
 
 END_C_DECLS
